finish removing http/connection
