+++ stardog/connection.py
@@ -6,7 +6,6 @@ import distutils.util
 
 from . import content_types as content_types
 from . import exceptions as exceptions
-from .http import connection as http_connection
 from .http import client
 
 
@@ -31,8 +30,6 @@ class Connection(object):
           >>> conn = Connection('db', endpoint='http://localhost:9999',
                                 username='admin', password='admin')
         """
-        self.conn = http_connection.Connection(database, endpoint, username,
-                                               password)
         self.client = client.Client(endpoint, database, username, password)
         self.transaction = None
 
@@ -76,7 +73,8 @@ class Connection(object):
               If already in a transaction
         """
         self._assert_not_in_transaction()
-        self.transaction = self.conn.begin(**kwargs)
+        r = self.client.post('/transaction/begin', params=kwargs)
+        self.transaction = r.text
         return self.transaction
 
     def rollback(self):
@@ -88,7 +86,7 @@ class Connection(object):
 
         """
         self._assert_in_transaction()
-        self.conn.rollback(self.transaction)
+        self.client.post('/transaction/rollback/{}'.format(self.transaction))
         self.transaction = None
 
     def commit(self):
@@ -99,7 +97,7 @@ class Connection(object):
             If currently not in a transaction
         """
         self._assert_in_transaction()
-        self.conn.commit(self.transaction)
+        self.client.post('/transaction/commit/{}'.format(self.transaction))
         self.transaction = None
 
     def add(self, content, graph_uri=None):
@@ -119,8 +117,14 @@ class Connection(object):
         self._assert_in_transaction()
 
         with content.data() as data:
-            self.conn.add(self.transaction, data, content.content_type,
-                          content.content_encoding, graph_uri)
+            self.client.post(
+                '/{}/add'.format(self.transaction),
+                params={'graph-uri': graph_uri},
+                headers={
+                    'Content-Type': content.content_type,
+                    'Content-Encoding': content.content_encoding
+                },
+                data=data)
 
     def remove(self, content, graph_uri=None):
         """Removes data from the database.
@@ -140,8 +144,14 @@ class Connection(object):
         self._assert_in_transaction()
 
         with content.data() as data:
-            self.conn.remove(self.transaction, data, content.content_type,
-                             content.content_encoding, graph_uri)
+            self.client.post(
+                '/{}/remove'.format(self.transaction),
+                params={'graph-uri': graph_uri},
+                headers={
+                    'Content-Type': content.content_type,
+                    'Content-Encoding': content.content_encoding
+                },
+                data=data)
 
     def clear(self, graph_uri=None):
         """Removes all data from the database or specific named graph.
@@ -163,7 +173,10 @@ class Connection(object):
           >>> conn.clear()
         """
         self._assert_in_transaction()
-        self.conn.clear(self.transaction, graph_uri)
+        self.client.post(
+            '/{}/clear'.format(self.transaction),
+            params={'graph-uri': graph_uri}
+        )
 
     def size(self, exact=False):
         """Database size.
@@ -174,7 +187,8 @@ class Connection(object):
         Returns:
           int: The number of elements in database
         """
-        return self.conn.size(exact=exact)
+        r = self.client.get('/size', params={'exact': exact})
+        return int(r.text)
 
     def export(self,
                content_type=content_types.TURTLE,
@@ -206,7 +220,14 @@ class Connection(object):
           >>> with conn.export(stream=True) as stream:
                 contents = ''.join(stream)
         """
-        db = self.conn.export(content_type, stream, chunk_size, graph_uri)
+        def _export():
+            with self.client.get(
+                    '/export', headers={'Accept': content_type},
+                    params={'graph-uri': graph_uri},
+                    stream=stream) as r:
+                yield r.iter_content(
+                    chunk_size=chunk_size) if stream else r.content
+        db = _export()
         return _nextcontext(db) if stream else next(db)
 
     def explain(self, query, base_uri=None):
@@ -219,7 +240,45 @@ class Connection(object):
         Returns:
          str: Query explanation
         """
-        return self.conn.explain(query, base_uri)
+        params = {'query': query, 'baseURI': base_uri}
+
+        r = self.client.post(
+            '/explain',
+            data=params,
+        )
+
+        return r.text
+
+    def __query(self,
+                query,
+                method,
+                content_type=None,
+                **kwargs):
+        txId = self.transaction
+        params = {
+            'query': query,
+            'baseURI': kwargs.get('base_uri'),
+            'limit': kwargs.get('limit'),
+            'offset': kwargs.get('offset'),
+            'timeout': kwargs.get('timeout'),
+            'reasoning': kwargs.get('reasoning')
+        }
+
+        # query bindings
+        bindings = kwargs.get('bindings', {})
+        for k, v in bindings.items():
+            params['${}'.format(k)] = v
+
+        url = '/{}/{}'.format(txId, method) if txId else '/{}'.format(method)
+
+        r = self.client.post(
+            url,
+            data=params,
+            headers={'Accept': content_type},
+        )
+
+        return r.json(
+        ) if content_type == content_types.SPARQL_JSON else r.content
 
     def select(self, query, content_type=content_types.SPARQL_JSON, **kwargs):
         """Executes a SPARQL select query.
@@ -251,8 +310,7 @@ class Connection(object):
 
           >>> conn.select('select * {?s ?p ?o}', bindings={'o': '<urn:a>'})
         """
-        return self.conn.query(
-            query, self.transaction, content_type=content_type, **kwargs)
+        return self.__query(query, 'query', content_type=content_type, **kwargs)
 
     def graph(self, query, content_type=content_types.TURTLE, **kwargs):
         """Executes a SPARQL graph query.
@@ -282,7 +340,7 @@ class Connection(object):
           >>> conn.graph('construct {?s ?p ?o} where {?s ?p ?o}',
                          bindings={'o': '<urn:a>'})
         """
-        return self.conn.query(query, self.transaction, content_type, **kwargs)
+        return self.__query(query, 'query', content_type, **kwargs)
 
     def paths(self, query, content_type=content_types.SPARQL_JSON, **kwargs):
         """Executes a SPARQL paths query.
@@ -310,7 +368,7 @@ class Connection(object):
           >>> conn.paths('paths start ?x = :subj end ?y = :obj via ?p',
                          reasoning=True)
         """
-        return self.conn.query(query, self.transaction, content_type, **kwargs)
+        return self.__query(query, 'query', content_type, **kwargs)
 
     def ask(self, query, **kwargs):
         """Executes a SPARQL ask query.
@@ -332,8 +390,7 @@ class Connection(object):
         Examples:
           >>> conn.ask('ask {:subj :pred :obj}', reasoning=True)
         """
-        r = self.conn.query(query, self.transaction, content_types.BOOLEAN,
-                            **kwargs)
+        r = self.__query(query, 'query', content_types.BOOLEAN, **kwargs)
         return bool(distutils.util.strtobool(r.decode()))
 
     def update(self, query, **kwargs):
@@ -353,7 +410,7 @@ class Connection(object):
         Examples:
           >>> conn.update('delete where {?s ?p ?o}')
         """
-        self.conn.update(query, self.transaction, **kwargs)
+        self.__query(query, 'update', None, **kwargs)
 
     def is_consistent(self, graph_uri=None):
         """Checks if the database or named graph is consistent wrt its schema.
@@ -365,7 +422,12 @@ class Connection(object):
         Returns:
           bool: Database consistency state
         """
-        return self.conn.is_consistent(graph_uri)
+        r = self.client.get(
+            '/reasoning/consistency',
+            params={'graph-uri': graph_uri},
+        )
+
+        return bool(distutils.util.strtobool(r.text))
 
     def explain_inference(self, content):
         """Explains the given inference results.
@@ -379,10 +441,21 @@ class Connection(object):
         Examples:
           >>> conn.explain_inference(File('inferences.ttl'))
         """
+        txId = self.transaction
+
         with content.data() as data:
-            return self.conn.explain_inference(data, content.content_type,
-                                               content.content_encoding,
-                                               self.transaction)
+            url = '/reasoning/{}/explain'.format(txId) if txId else '/reasoning/explain'
+
+            r = self.client.post(
+                url,
+                data=data,
+                headers={
+                    'Content-Type': content.content_type,
+                    'Content-Encoding': content.content_encoding
+                },
+            )
+
+            return r.json()['proofs']
 
     def explain_inconsistency(self, graph_uri=None):
         """Explains why the database or a named graph is inconsistent.
@@ -394,8 +467,16 @@ class Connection(object):
         Returns:
           dict: Explanation results
         """
-        return self.conn.explain_inconsistency(self.transaction,
-                                               graph_uri)
+        txId = self.transaction
+        url = '/reasoning/{}/explain/inconsistency'.format(
+            txId) if txId else '/reasoning/explain/inconsistency'
+
+        r = self.client.get(
+            url,
+            params={'graph-uri': graph_uri},
+        )
+
+        return r.json()['proofs']
 
     def _assert_not_in_transaction(self):
         if self.transaction:
@@ -405,11 +486,16 @@ class Connection(object):
         if not self.transaction:
             raise exceptions.TransactionException('Not in a transaction')
 
+    def close(self):
+        """Close the underlying HTTP connection.
+        """
+        self.client.close()
+
     def __enter__(self):
         return self
 
     def __exit__(self, *args):
-        self.conn.__exit__(*args)
+        self.close()
 
 
 class Docs(object):
